+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/entity/Examination.java
@@ -1,0 +1,45 @@
+package com.ivanovp.medical_record.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import jakarta.persistence.*;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "examinations")
+@Getter
+@Setter
+@NoArgsConstructor
+public class Examination {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "exam_date", nullable = false)
+    private LocalDateTime examDate;
+
+    @Column(nullable = true)
+    private String treatment;
+
+    @Column(nullable = false)
+    private BigDecimal price;
+
+    @Column(name = "paid_by_nzok", nullable = false)
+    private boolean paidByNzok;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "doctor_id", nullable = false)
+    private Doctor doctor;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "patient_id", nullable = false)
+    private Patient patient;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "diagnosis_id", nullable = true)
+    private Diagnosis diagnosis;
+}