+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/entity/Patient.java
@@ -1,0 +1,36 @@
+package com.ivanovp.medical_record.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import jakarta.persistence.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@Entity
+@Table(name = "patients")
+public class Patient {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "egn", unique = true, nullable = false)
+    private String egn;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "is_insured", nullable = false, columnDefinition = "BIT DEFAULT 1")
+    private boolean isInsured = true;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "gp_id", nullable = true)
+    private Doctor gp;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", unique = true, nullable = true)
+    private User user;
+}