+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/entity/SickLeave.java
@@ -1,0 +1,30 @@
+package com.ivanovp.medical_record.entity;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import jakarta.persistence.*;
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "sick_leaves")
+@Getter
+@Setter
+@NoArgsConstructor
+public class SickLeave {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "start_date", nullable = false)
+    private LocalDate startDate;
+
+    @Column(name = "duration_days", nullable = false)
+    private int durationDays;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "examination_id", unique = true, nullable = false)
+    private Examination examination;
+}