+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/repository/DiagnosisRepository.java
@@ -1,0 +1,8 @@
+package com.ivanovp.medical_record.repository;
+
+import com.ivanovp.medical_record.entity.Diagnosis;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DiagnosisRepository extends JpaRepository<Diagnosis, Long> {
+    boolean existsByCode(String code);
+}