+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/repository/SpecialtyRepository.java
@@ -1,0 +1,8 @@
+package com.ivanovp.medical_record.repository;
+
+import com.ivanovp.medical_record.entity.Specialty;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SpecialtyRepository extends JpaRepository<Specialty, Long> {
+    boolean existsByName(String name);
+}