+++ medical-record/medical-record/src/main/java/com/ivanovp/medical_record/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.ivanovp.medical_record.repository;
+
+import com.ivanovp.medical_record.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+}